requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.5
path==17.0.0
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
